requirements.txt
keras-cv
keras
tensorflow
datasets
transformers
evaluate
tf-keras
pillow
matplotlib
scikit-learn
jax